Split out drone scanning
