+++ Assets/Scriptable Objects/State Machines/Drone/Parameters/ScanButtonPressed.asset.meta
