+++ Assets/Scripts/Drone/Scanning.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1a1a5e9044ef95848bbfad4c2f1efe3f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
