+++ Assets/Scripts/Drone/Scanning/DroneScannableLocator.cs
