+++ Assets/Scripts/Drone/Scanning/DroneScanning.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 527c8e605aff57d46b9df0ad42c7d111
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
